RO 3.2.1
Changed RO.Wdg.Sound to raise RuntimeError if the specified sound file does not exist.
Unfortunately there is no warning if the file exists but is not in a usable format
(I don't know how to get pygame.mixer.Sound to check; at one time I think it raised
an exception but it certainly does not do so now).